Rename fbschemas.* to FlatbufferMessage.*
Updates #58
